+++ @xen-orchestra/lite/src/components/host/HostHeader.vue
@@ -20,6 +20,22 @@
         </template>
         <HostPowerStateActions :host />
       </MenuList>
+
+      <MenuList placement="bottom-end">
+        <template #trigger="{ open }">
+          <UiButtonIcon
+            v-tooltip="{
+              placement: 'left',
+              content: t('more-actions'),
+            }"
+            icon="action:more-actions"
+            accent="brand"
+            size="medium"
+            @click="open($event)"
+          />
+        </template>
+        <HostMoreActions :host />
+      </MenuList>
     </template>
   </UiHeadBar>
 </template>
@@ -27,12 +43,14 @@
 <script lang="ts" setup>
 import type { XenApiHost } from '@/libs/xen-api/xen-api.types.ts'
 import HostPowerStateActions from '@/modules/host/components/actions/HostPowerStateActions.vue'
+import HostMoreActions from '@/modules/host/components/HostMoreActions.vue'
 import { useHostUtils } from '@/modules/host/composables/host-utils.composable.ts'
 import { useHostMetricsStore } from '@/stores/xen-api/host-metrics.store.ts'
 import { usePoolStore } from '@/stores/xen-api/pool.store.ts'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
 import MenuList from '@core/components/menu/MenuList.vue'
 import VtsObjectIcon from '@core/components/object-icon/VtsObjectIcon.vue'
+import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiDropdownButton from '@core/components/ui/dropdown-button/UiDropdownButton.vue'
 import UiHeadBar from '@core/components/ui/head-bar/UiHeadBar.vue'
 import { vTooltip } from '@core/directives/tooltip.directive.ts'